Debug #2

+++ all_models_source_target.py
@@ -57,8 +57,8 @@ def binary_noise(generator, ae: bool):
     def noise_gen():
         for X, y in generator:
             for n, matr in enumerate(X):
-                x_indices = randint(0, matr.shape[0], size=BIN_DROP * matr.shape[0])
-                y_indices = randint(0, matr.shape[1], size=BIN_DROP * matr.shape[1])
+                x_indices = randint(0, matr.shape[0], size=int(BIN_DROP * matr.shape[0]))
+                y_indices = randint(0, matr.shape[1], size=int(BIN_DROP * matr.shape[1]))
                 X[n, x_indices, y_indices] = 0
             if ae:
                 yield X, X
